requirements.txt
textblob==0.17.1
python-dateutil==2.8.2
requests==2.31.0
pytest==7.4.0